move util functions to a single file
